Add tests for getSelectors function
This allows certainty that selectors are being unwrapped as expected. By
following the 'black box' style used in other tests it can lead to false
positives. For example:

```css
.Component {
  .Component-descendant {
    color: green;
  }
}
```

In the above code what should be linted is
`.Component .Component-descendant` but it's quite possible due to an
error that just `.Component` is tested. If this happens it will still
pass the black box test as `.Component` is valid. This can lead to hard
to find bugs with unwrapping nesting selectors.